+++ tests/test_007_flowsheet.py
@@ -1,9 +1,6 @@
-import pytest
-
 from elphick.geomet import Stream
 from elphick.geomet.flowsheet import Flowsheet
-from elphick.geomet.base import MC
-from elphick.geomet.operation import NodeType, Operation
+from elphick.geomet.operation import Operation
 from fixtures import sample_data
 
 
@@ -32,7 +29,6 @@ def test_flowsheet_init(sample_data):
 
 def test_solve(sample_data):
     # Create a new Flowsheet object
-    fs = Flowsheet()
     obj_strm: Stream = Stream(sample_data, name='Feed')
     obj_strm_1, obj_strm_2 = obj_strm.split(0.4, name_1='stream 1', name_2='stream 2')
     fs: Flowsheet = Flowsheet.from_objects([obj_strm, obj_strm_1, obj_strm_2])
@@ -54,12 +50,56 @@ def test_solve(sample_data):
 
 def test_query(sample_data):
     # Create a new Flowsheet object
-    fs = Flowsheet()
     obj_strm: Stream = Stream(sample_data, name='Feed')
     obj_strm_1, obj_strm_2 = obj_strm.split(0.4, name_1='stream 1', name_2='stream 2')
     fs: Flowsheet = Flowsheet.from_objects([obj_strm, obj_strm_1, obj_strm_2])
 
-    # Call the query method
-    fs_reduced: Flowsheet = fs.query(query='Fe>57')
+    # Call the query method with inplace=False
+    fs_reduced: Flowsheet = fs.query(query='Fe>58', inplace=False)
 
-    # TODO: enhance this test
+    # Check that the original flowsheet remains unmutated
+    assert fs.get_input_streams()[0].data.equals(obj_strm.data)
+    assert fs.get_output_streams()[0].data.equals(obj_strm_1.data)
+    assert fs.get_output_streams()[1].data.equals(obj_strm_2.data)
+
+    # Check that the filtered flowsheet has the correct data
+    assert not fs_reduced.get_input_streams()[0].data.equals(obj_strm.data)
+    assert not fs_reduced.get_output_streams()[0].data.equals(obj_strm_1.data)
+    assert not fs_reduced.get_output_streams()[1].data.equals(obj_strm_2.data)
+
+    # Call the query method with inplace=True
+    fs.query(query='Fe>58', inplace=True)
+
+    # Check that the original flowsheet is now mutated, matching the fs_reduced
+    assert fs.get_input_streams()[0].data.equals(fs_reduced.get_input_streams()[0].data)
+    assert fs.get_output_streams()[0].data.equals(fs_reduced.get_output_streams()[0].data)
+    assert fs.get_output_streams()[1].data.equals(fs_reduced.get_output_streams()[1].data)
+
+
+def test_filter_by_index(sample_data):
+    # Create a new Flowsheet object
+    obj_strm: Stream = Stream(sample_data, name='Feed')
+    obj_strm_1, obj_strm_2 = obj_strm.split(0.4, name_1='stream 1', name_2='stream 2')
+    fs: Flowsheet = Flowsheet.from_objects([obj_strm, obj_strm_1, obj_strm_2])
+
+    # Call the filter method with inplace=False
+    index = obj_strm.data.index[1:]
+    fs_reduced: Flowsheet = fs.filter_by_index(index=index, inplace=False)
+
+    # Check that the original flowsheet remains unmutated
+    assert fs.get_input_streams()[0].data.equals(obj_strm.data)
+    assert fs.get_output_streams()[0].data.equals(obj_strm_1.data)
+    assert fs.get_output_streams()[1].data.equals(obj_strm_2.data)
+
+    # Check that the filtered flowsheet has the correct data
+    assert not fs_reduced.get_input_streams()[0].data.equals(obj_strm.data)
+    assert not fs_reduced.get_output_streams()[0].data.equals(obj_strm_1.data)
+    assert not fs_reduced.get_output_streams()[1].data.equals(obj_strm_2.data)
+
+    # Call the filter method with inplace=True
+    fs.filter_by_index(index=index, inplace=True)
+
+    # Check that the original flowsheet is now mutated, matching the fs_reduced
+    assert fs.get_input_streams()[0].data.equals(fs_reduced.get_input_streams()[0].data)
+    assert fs.get_output_streams()[0].data.equals(fs_reduced.get_output_streams()[0].data)
+    assert fs.get_output_streams()[1].data.equals(fs_reduced.get_output_streams()[1].data)