+++ CHANGELOG.rst
@@ -1,3 +1,21 @@
+Geomet 0.2.1 (2024-08-17)
+=========================
+
+Feature
+-------
+
+- Added query and filter_by_index methods to Flowsheet.  Added example. (#18)
+
+
+Geomet 0.2.0 (2024-08-10)
+=========================
+
+Feature
+-------
+
+- Added ability to create a Flowsheet from file, from_yaml, from_json (#14)
+
+
 Geomet 0.1.1 (2024-06-19)
 =========================